Project-8_OpenCV_Face_Detection/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.2
importlib-metadata==4.6.1
imutils==0.5.4
numpy==1.21.0 #check newest version
opencv-python==4.5.3.56
Pillow==8.3.1
PyJWT==1.7.1
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==1.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
sendgrid==6.10.0


